fix: address PR review feedback
Implemented changes from review comments:
- [Gemini] Moved Contact interface to src/app/core/models/contact.model.ts
