+++ src/app/core/models/contact.model.ts
@@ -1,0 +1,7 @@
+export interface Contact {
+  id: string; // email or uid
+  email: string;
+  displayName: string;
+  photoURL?: string;
+  source: 'workspace' | 'google-contacts' | 'google-directory' | 'default-domain' | 'cached';
+}
